Remove SettingsImporterImpl
